calibration file added
Creates a file with calibrated speeds for quicker testing. If TESTING is true, then the robot is in testing mode. If TEST_WRITE is true, the robot will calibrate and write the results to the file. If TEST_WRITE is false, the robot will pull from the calibration file. If TESTING is false, the robot will run normally.